refactor(wave3-step2): scaffold monitor/trace split behind stable facades (#338)
* refactor(wave3-step2): add compile-safe monitor/trace split scaffolds

* refactor(wave3-step2): mechanically split monitor/trace impl behind facades

* fix(wave3-step2): preserve trace diagnostic wording

* docs(wave3-step2): add hard boundary gates and reviewer artifacts

* docs(wave3-step2): tighten unsafe and output containment gates

* fix(wave3-step2): borrow kill mode in log path to avoid move

* fix(wave3-step2): remove duplicate linux cfg attribute
