+++ crates/assay-cli/src/cli/commands/monitor_next/errors.rs
@@ -1,0 +1,5 @@
+//! Monitor split error helpers and mappings.
+//!
+//! Boundary:
+//! - typed mapping helpers only
+//! - no runtime side effects
