+++ crates/assay-cli/src/cli/commands/monitor_next/tests.rs
@@ -1,0 +1,6 @@
+#![cfg(test)]
+
+//! No tests moved in Step2 Commit A.
+//! Step1 monitor contract tests remain in:
+//! `crates/assay-cli/src/cli/commands/monitor.rs`.
+//! See move-map/checklist artifacts in Commit C.
