+++ crates/assay-core/src/providers/trace_next/tests.rs
@@ -1,0 +1,6 @@
+#![cfg(test)]
+
+//! No tests moved in Step2 Commit A.
+//! Step1 trace provider contract tests remain in:
+//! `crates/assay-core/src/providers/trace.rs`.
+//! See move-map/checklist artifacts in Commit C.
